requirements.txt
boto3>=1.26.0
# openai-whisper>=20231117
torch>=2.0.0
transformers>=4.30.0
PyYAML>=6.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.20.0 